requirements.txt
Flask==2.3.2
Flask-MySQLdb==2.0.0
mysqlclient==2.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
pandas==2.0.3